requirements.txt
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
imageio==2.10.1
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.3
opencv-contrib-python-headless==4.5.4.58
Pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.11.2
Werkzeug==2.0.2
